pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "mtp-gateway"
dynamic = ["version"]
description = "Production-grade gateway bridging legacy PLCs to MTP-compliant OPC UA interface"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "MTP Gateway Team" }
]
keywords = ["mtp", "opcua", "plc", "modbus", "automation", "vdi2658"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
]

dependencies = [
    "asyncua>=1.0.0",
    "pymodbus>=3.6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=23.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "defusedxml>=0.7.1",
    "anyio>=4.0.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
]

[project.optional-dependencies]
s7 = ["python-snap7>=1.3"]
eip = ["pycomm3>=1.2.0"]
webui = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.25",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "jsonschema<4.18",
    "bandit>=1.7.6",
    "pip-audit>=2.7.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "types-defusedxml>=0.7.0.20240125",
    "types-hvac>=0.1.0.20240125",
    "types-PyYAML>=6.0.0",
]
all = ["mtp-gateway[s7,eip,webui,dev]"]

[project.scripts]
mtp-gateway = "mtp_gateway.cli.app:app"

[project.urls]
Documentation = "https://github.com/example/mtp-gateway#readme"
Issues = "https://github.com/example/mtp-gateway/issues"
Source = "https://github.com/example/mtp-gateway"

[tool.hatch.build.targets.wheel]
packages = ["src/mtp_gateway"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

# Version from git tags (hatch-vcs)
[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"  # Clean versions for PyPI

[tool.hatch.build.hooks.vcs]
version-file = "src/mtp_gateway/_version.py"

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]
exclude = ["src/mtp_gateway/_version.py"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # pylint
    "RUF",    # Ruff-specific
    "ASYNC",  # flake8-async
]
ignore = [
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value comparison
]

[tool.ruff.lint.isort]
known-first-party = ["mtp_gateway"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001", "PLR2004"]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "mtp_gateway._version",
    "hvac.*",
    "pymodbus.*",
    "snap7.*",
    "pycomm3.*",
    "asyncua.*",
    "jose.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "-p",
    "no:schemathesis",
]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (may require external services)",
    "contract: Contract tests (manifest/server consistency)",
    "slow: Slow tests (>1s)",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:asyncua.*",
    "ignore::DeprecationWarning:dateutil.*",
    "ignore::DeprecationWarning:jsonschema.*",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/mtp_gateway"]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
fail_under = 60
show_missing = true

# Bandit security configuration
[tool.bandit]
targets = ["src"]
exclude_dirs = ["tests"]
skips = ["B101"]  # Skip assert warnings (used in tests)
